# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: graphRAG Console API
  version: "1.1"
  description: |
    Frozen OpenAPI contract for the graphRAG Console backend (DESIGN §15, DR-002).
    `core`/`api`/`web` all build against this document; the FastAPI app (BA0) must
    generate a spec that matches it.

    ## Conventions (frozen)
    - **Success envelope** — every JSON success body is
      `{ "data": <payload>, "meta": { "request_id", "build_id", "elapsed_ms" } }`.
    - **Error envelope** — every error body is
      `{ "error": { "code": <ErrorCode>, "message", "details", "request_id" } }`.
      Status mapping: 400 `VALIDATION_ERROR`/`QUERY_UNSAFE` · 404 `*_NOT_FOUND` ·
      409 `IDEMPOTENCY_CONFLICT`/`JOB_CONFLICT`/`BUILD_NOT_READY`/`NO_ACTIVE_BUILD` ·
      429 `RATE_LIMITED` · 503 `STORE_UNAVAILABLE` · 504 `QUERY_TIMEOUT` · 500 `INTERNAL`.
    - **Pagination** — list endpoints take `?limit=&cursor=` (opaque cursor) and return
      `meta.next_cursor` (`null` on the last page).
    - **Sorting / filtering** — `?sort=field:asc|desc` and `?filter[field]=value`.
    - **Long operations** — return `202` with `{data: {job_id, status}}`; progress via
      `GET /jobs/{job_id}/events` (SSE, events `job.update|job.done|job.failed`).
    - **Idempotency** — every POST outside `/query/*` accepts an `Idempotency-Key`
      header. Same key + same request hash → the stored original response is replayed;
      same key + different request hash → `409 IDEMPOTENCY_CONFLICT`. Keys expire after
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /projects/{project}/clean/preview:
    parameters:
      - $ref: "#/components/parameters/ProjectPath"
    post:
      operationId: previewClean
      tags: [clean]
      summary: Preview chunking (§10.2 抽樣預覽) — a pure function; nothing is persisted
      description: |
        Chunk ONE source with the given parameters and return the chunks that a build
        WOULD produce, without writing anything. The source is exactly one of
        `document_id` (that document's raw text, read from the ACTIVE build) or `text`
        (chunk the given text — usable before any build exists). Omitted `max_chars` /
        `overlap` fall back to the project's configured chunking values, then to the
        engine defaults, so the no-parameter preview shows what a build would actually do.
        The pair must satisfy `0 <= overlap < max_chars` — a relation two independent
        numeric bounds cannot express in JSON Schema, so it is enforced at runtime
        (violations answer 400 VALIDATION_ERROR).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CleanPreviewRequest"
      responses:
        "200":
          description: |
            The chunks these parameters would produce. meta.build_id names the active
            build the document was read from; null for the `text` source (no build
            involved).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanPreviewResponse"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Auth placeholder (DESIGN §23) — swap the implementation, keep the scheme.
  parameters:
    ProjectPath:
      name: project
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Error:
      description: |
        Error envelope. Status → code mapping: 400 `VALIDATION_ERROR`/`QUERY_UNSAFE` ·
        404 `PROJECT_NOT_FOUND`/`BUILD_NOT_FOUND`/`JOB_NOT_FOUND` · 409
        `IDEMPOTENCY_CONFLICT`/`JOB_CONFLICT`/`BUILD_NOT_READY`/`NO_ACTIVE_BUILD` ·
        429 `RATE_LIMITED` · 503 `STORE_UNAVAILABLE` · 504 `QUERY_TIMEOUT` · 500 `INTERNAL`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: PROJECT_NOT_FOUND
              message: project 'acme' does not exist
              details: null
              request_id: 6f1c2a4e-9b0d-4c39-9a55-1f2f3a4b5c6d
...
  schemas:
...
    Meta:
      type: object
      description: Success-envelope metadata (DESIGN §15).
      required: [request_id, build_id, elapsed_ms]
      properties:
        request_id:
          type: string
          format: uuid
        build_id:
          description: Build that served the request; null when no build was involved.
          type: ["string", "null"]
          format: uuid
        elapsed_ms:
          type: integer
          minimum: 0
...
    ErrorCode:
      type: string
      description: Frozen error-code enum (DESIGN §27.2; additive-only).
      enum:
        - PROJECT_NOT_FOUND
        - BUILD_NOT_FOUND
        - BUILD_NOT_READY
        - NO_ACTIVE_BUILD
        - VALIDATION_ERROR
        - JOB_NOT_FOUND
        - JOB_CONFLICT
        - IDEMPOTENCY_CONFLICT
        - QUERY_UNSAFE
        - QUERY_TIMEOUT
        - STORE_UNAVAILABLE
        - RATE_LIMITED
        - INTERNAL
    Error:
      type: object
      description: |
        The full frozen shape is always present — details is null rather than absent,
        so error consumers never branch on missing fields.
      required: [code, message, details, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          description: Machine-readable context (field errors, guardrail verdicts, …).
          type: ["object", "null"]
          additionalProperties: true
        request_id:
          type: string
          format: uuid
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
...
    CleanPreviewByDocument:
      type: object
      description: Chunk a document's raw text, read from the ACTIVE build.
      additionalProperties: false
      required: [document_id]
      properties:
        max_chars:
          description: Window size; omitted → project config, then engine default.
          type: integer
          minimum: 1
        overlap:
          description: Window overlap; omitted → project config, then engine default.
          type: integer
          minimum: 0
        document_id:
          type: string
          format: uuid
        text: false
    CleanPreviewByText:
      type: object
      description: Chunk the given text directly — works before any build exists.
      additionalProperties: false
      required: [text]
      properties:
        max_chars:
          description: Window size; omitted → project config, then engine default.
          type: integer
          minimum: 1
        overlap:
          description: Window overlap; omitted → project config, then engine default.
          type: integer
          minimum: 0
        text:
          type: string
          minLength: 1
        document_id: false
    CleanPreviewRequest:
      description: |
        Exactly ONE source: document_id or text. The two variants are concrete
        object schemas (each closed with additionalProperties: false), and each
        declares the OTHER source with the `false` schema — so codegen emits it
        as `?: never` and the union is EXACT, not merely structural: a payload
        naming both sources fails to type even when built in a variable (excess-
        property checks only cover literals). Both sources, or neither, is 400
        at the API either way.
      oneOf:
        - $ref: "#/components/schemas/CleanPreviewByDocument"
        - $ref: "#/components/schemas/CleanPreviewByText"
    CleanPreviewChunk:
      type: object
      description: |
        One previewed chunk. Computed on the fly and never stored, so it carries no
        id/document_id/build_id — offsets satisfy raw[start_offset:end_offset] == text
        exactly (the same invariant stored chunks freeze for §27.4 evidence spans).
      required: [ordinal, text, start_offset, end_offset, token_count]
      properties:
        ordinal:
          type: integer
          minimum: 0
        text:
          type: string
          minLength: 1
        start_offset:
          type: integer
          minimum: 0
        end_offset:
          description: Exclusive end, per Python slice semantics.
          type: integer
          minimum: 1
        token_count:
          type: integer
          minimum: 1
    CleanPreviewResult:
      type: object
      required: [chunks]
      properties:
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/CleanPreviewChunk"
    CleanPreviewResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/CleanPreviewResult"
        meta:
          $ref: "#/components/schemas/Meta"
...